Updated read store tests for MSSQL
