+++ Source/EventFlow.MsSql.Tests/IntegrationTestsForReadStore/QueryHandlers/MsSqlThingyGetMessagesQueryHandler.cs
@@ -1,0 +1,60 @@
+﻿// The MIT License (MIT)
+// 
+// Copyright (c) 2015 Rasmus Mikkelsen
+// Copyright (c) 2015 eBay Software Foundation
+// https://github.com/rasmus/EventFlow
+// 
+// Permission is hereby granted, free of charge, to any person obtaining a copy of
+// this software and associated documentation files (the "Software"), to deal in
+// the Software without restriction, including without limitation the rights to
+// use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+// the Software, and to permit persons to whom the Software is furnished to do so,
+// subject to the following conditions:
+// 
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+// FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+// COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+// IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+// CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+// 
+
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using EventFlow.Core;
+using EventFlow.MsSql.Tests.IntegrationTestsForReadStore.ReadModels;
+using EventFlow.Queries;
+using EventFlow.TestHelpers.Aggregates.Entities;
+using EventFlow.TestHelpers.Aggregates.Queries;
+
+namespace EventFlow.MsSql.Tests.IntegrationTestsForReadStore.QueryHandlers
+{
+    public class MsSqlThingyGetMessagesQueryHandler : IQueryHandler<ThingyGetMessagesQuery, IReadOnlyCollection<ThingyMessage>>
+    {
+        private readonly IMsSqlConnection _msSqlConnection;
+
+        public MsSqlThingyGetMessagesQueryHandler(
+            IMsSqlConnection msSqlConnection)
+        {
+            _msSqlConnection = msSqlConnection;
+        }
+
+        public async Task<IReadOnlyCollection<ThingyMessage>> ExecuteQueryAsync(ThingyGetMessagesQuery query, CancellationToken cancellationToken)
+        {
+            var readModels = await _msSqlConnection.QueryAsync<MsSqlThingyMessageReadModel>(
+                Label.Named("mssql-fetch-thingy-message-read-model"),
+                cancellationToken,
+                "SELECT * FROM [ReadModel-ThingyMessage] WHERE ThingyId = @ThingyId",
+                new { ThingyId = query.ThingyId.Value })
+                .ConfigureAwait(false);
+            return readModels
+                .Select(rm => rm.ToThingyMessage())
+                .ToList();
+        }
+    }
+}
