+++ Source/EventFlow.MsSql.Tests/IntegrationTestsForReadStore/Scripts/0001 - Create table ReadModel-ThingyAggregate.sql
